Expose Next Game DateTime as Attribute
This adds a new attribute, "Next game datetime," which exposes the date and time of the next game in a single attribute. This is necessary if you want to have automations that rely on the start time. 

Probably not being defined in the best place; maybe there's a better place for it to be placed?